+++ UDP Server/src/model/Bullet.java
@@ -2,9 +2,9 @@ package model;
 
 public class Bullet extends GameObject{
 	
-	public int direction;
+	public short direction;
 	
-	public Bullet(float x, float y, int direction) {
+	public Bullet(float x, float y, short direction) {
 		super(x, y);
 		this.direction = direction;
 	}
